requirements.txt
crawl4ai
firecrawl-py
tavily_python
python-dotenv
agno
streamlit